backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.9
redis>=5.0.3
celery>=5.3.6
pandas>=2.2.2
numpy>=1.26.4
sentence-transformers>=2.6.1
python-multipart>=0.0.9
slowapi>=0.1.9
asyncpg>=0.29.0
firebase-admin>=6.5.0
loguru>=0.7.2
google-generativeai>=0.8.0
httpx>=0.27.0
